added switching functions
the switching functions are implemented at the moment as enums. I am not sure if this is the best way to distinguish the different switching functions. Furthermore, we have to very similiar functions for the switching function value and the value of derivate. It might be better to use only kind of definition and distinguish if we want to compute the function value or the derivative.