added coordinations to Atom
added a field that holds the xyz-coordinates of an atom to the Atom struct. this makes it possible to compute the gamma matrix without a computation of the distance matrix and directions matrix before. thee coordinates are stored as a nalgebra Vector3, since this should be much faster than an Array1 from ndarray since it is statically allocated on the stack.