refactured the fmo scc code
the individual contributions to the FMO energy are now computed in seperate functions. this makes it then easier to compute the numerical gradients for each part. the wrapper for the numerical gradients would be the next step. the current problem is, that the atoms are stored in the SuperSystem, Monomer and Pair as their own data. this makes it complex and expensive to update the positions of the atoms. self referential structs are not possible because at the moment of updating the positions there can no references to the atoms. so the idea was to store the atoms only in the SuperSystem and give slices to the monomer and pair functions. but in the SCC calculations of the monomer there is also a conflict of the mutable borrow of the monomer and of the atoms. this has to be solved, by designing a better way to store the atoms.